chore: refactor theme
